+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ## v2.14.1
 ### Changed
 - Updated Inference Extra Networks (Lora / LyCORIS) base model filtering to consider SDXL variations (e.g., Noob AI / Illustrious) as compatible, unrecognized models or models with no base model will be considered compatible.
+- Changed hotkey for Inference prompt weight adjustment to (`⌘+Up`/`⌘+Down`) on macOS
 ### Fixed
 - Fixed Inference hotkey weight adjustment multi-line behavior, now works as expected like the first line.
 - Fixed updates to versions with commit hash version parts not being recognized when the current version has no commit hash part.